core/Cuda: Make floating_point_wrapper backend agnostic
